+++ index.html
@@ -6,8 +6,7 @@
     <meta name="description" content="Log in to start your journey!" />
     <link rel="icon" href="/images/tab-icon.webp">
     <link href="/dist/css/styles.css" rel="stylesheet" />
-    <script type="module" src="/js/components/register.mjs"></script>
-    <script type="module" src="/js/components/login.mjs"></script>
+    <script type="module" src="/js/index.mjs"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
