+++ js/components/login.mjs
@@ -1,12 +1,7 @@
-import { API_BASE_URL } from "../index.mjs";
 import { apiFetch } from "./apiFetch.mjs";
-const API_SOCIAL_LOGIN_PATH = "/social/auth/login";
-const API_SOCIAL_LOGIN_URL = API_BASE_URL + API_SOCIAL_LOGIN_PATH;
+import { API_SOCIAL_LOGIN_URL } from "./urls.mjs";
 
-const loginForm = document.querySelector(".log-in-form");
-const loginButton = document.querySelector(".log-in-button");
-
-function login(event) {
+export function login(event) {
   event.preventDefault();
   const loginEmail = document.querySelector(".log-in-email").value;
   const loginPassword = document.querySelector(".log-in-password").value;
@@ -30,5 +25,3 @@ function setToken(json) {
   localStorage.setItem("username", json.name)
   window.location.href = "/feed.html";
 }
-
-loginButton.addEventListener("click", login);
