+++ js/components/makePost.mjs
@@ -1,7 +1,5 @@
-import { API_BASE_URL } from "../index.mjs";
+import { API_SOCIAL_POSTS_URL } from "./urls.mjs";
 import { apiFetch } from "./apiFetch.mjs";
-const API_SOCIAL_POST_PATH = "/social/posts";
-const API_SOCIAL_POST_URL = `${API_BASE_URL}${API_SOCIAL_POST_PATH}`;
 const accessToken = localStorage.getItem("userToken");
 const submitPostInput = document.querySelector(".post-text");
 
@@ -19,8 +17,7 @@ export function makePost(e) {
     }),
   }
 
-  console.log(postInfo);
-  apiFetch(API_SOCIAL_POST_URL, postInfo, logging);
+  apiFetch(API_SOCIAL_POSTS_URL, postInfo, logging);
 }
 
 function logging (file) {
