+++ js/components/urls.mjs
@@ -1,0 +1,10 @@
+const API_BASE_URL = "https://api.noroff.dev/api/v1";
+
+const API_SOCIAL_LOGIN_PATH = "/social/auth/login";
+export const API_SOCIAL_LOGIN_URL = `${API_BASE_URL}${API_SOCIAL_LOGIN_PATH}`;
+
+const API_SOCIAL_REGISTER_PATH = "/social/auth/register";
+export const API_SOCIAL_REGISTER_URL = `${API_BASE_URL}${API_SOCIAL_REGISTER_PATH}`;
+
+const API_SOCIAL_POSTS_PATH = "/social/posts";
+export const API_SOCIAL_POSTS_URL = `${API_BASE_URL}${API_SOCIAL_POSTS_PATH}`;
