+++ js/index.mjs
@@ -1,1 +1,8 @@
-export const API_BASE_URL = "https://api.noroff.dev/api/v1";
+import { login } from "./components/login.mjs";
+import { registerUser } from "./components/register.mjs";
+
+const loginButton = document.querySelector(".log-in-button");
+const submitRegister = document.querySelector(".submit-register");
+
+loginButton.addEventListener("click", login);
+submitRegister.addEventListener("click", registerUser);
